requirements.txt
boto3==1.21.28
Django==4.0.3
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn
pillow
python-dotenv
psycopg2-binary
kavenegar
django-cors-headers
pydantic
drf-yasg